Make the Hero component more flexible
