+++ src/views/Homepage.tsx
@@ -1,4 +1,6 @@
 import React from "react";
+import Typography from "@material-ui/core/Typography";
+import Button from "@material-ui/core/Button";
 
 import Navbar from "../components/Navbar";
 import Hero from "../components/Hero";
@@ -9,12 +11,42 @@ import Subscribe from "../components/Subscribe";
 import mockData from "../mockData";
 
 function Homepage() {
+  const preventDefault = (event: React.MouseEvent) => event.preventDefault();
+
   return (
     <>
       <Navbar />
 
       <main>
-        <Hero />
+        <Hero
+          title="Welcome to The Bridge!"
+          cloudinaryImageTitle="/the-bridge/pexels-tope-a-asokere-4237407"
+        >
+          <Typography variant="body1">
+            Making international philanthropy easier by connecting donors with
+            social impact nonprofit organizations across Nigeria.
+          </Typography>
+
+          <Button
+            href="#"
+            onClick={preventDefault}
+            color="inherit"
+            variant="contained"
+            component="a"
+          >
+            Browse organizations
+          </Button>
+
+          <Button
+            href="#"
+            onClick={preventDefault}
+            color="inherit"
+            variant="contained"
+            component="a"
+          >
+            Learn more
+          </Button>
+        </Hero>
         <FeaturedOrgList data={mockData.orgs} />
         <OrgList data={mockData.orgs} />
       </main>
